Add synthetic images similar to synthetic flat tensor experiment

+++ datasets/synthetic/synthetic_im.py
@@ -1,19 +1,17 @@
 import os
 import random
 import shutil
-from typing import Callable, Optional
+from typing import Callable
 
-import torch
 from PIL import Image
 from torch.utils.data import Dataset
 from torchvision.transforms import ToTensor
 
-from datasets.im_generators import dummy_gen
 from utils.file import change_working_dir
 
 
 class Synthetic(Dataset):
-    def __init__(self, mode: str, generator: Callable[[], Image.Image], num: int, regen=False):
+    def __init__(self, mode: str, generator: Callable[[], Image.Image], num: int, regen=False, keep_in_ram=True):
         """
         Load or generate a set of synthetic images.
 
@@ -29,29 +27,32 @@ class Synthetic(Dataset):
         self.num = num
         change_working_dir()
         self.folder = f"data/{mode}"
+        self.keep_in_ram = keep_in_ram
 
-        if regen and os.path.isdir(self.folder):
-            shutil.rmtree(self.folder)
+        if self.keep_in_ram:
+            print(f"Generating {self.num} tensors")
+            self.imgs = [generator() for _ in range(self.num)]
+        else:
+            if regen and os.path.isdir(self.folder):
+                shutil.rmtree(self.folder)
 
-        os.makedirs(self.folder, exist_ok=True)
+            os.makedirs(self.folder, exist_ok=True)
 
-        existing_imgs = len(os.listdir(self.folder))
+            existing_imgs = len(os.listdir(self.folder))
 
-        if self.num > existing_imgs:
-            for i in range(existing_imgs, self.num):
-                im = generator()
-                im.save(os.path.join(self.folder, f"{i}.png"))
+            if self.num > existing_imgs:
+                for i in range(existing_imgs, self.num):
+                    im = generator()
+                    im.save(os.path.join(self.folder, f"{i}.png"))
 
-        existing_imgs = max(self.num, existing_imgs)
+            existing_imgs = max(self.num, existing_imgs)
 
-        self.imgs = random.sample(range(existing_imgs), self.num)
+            self.imgs = random.sample(range(existing_imgs), self.num)
 
     def __getitem__(self, index):
+        if self.keep_in_ram:
+            return self.imgs[index]
         return ToTensor()(Image.open(f"{self.imgs[index]}.png"))
 
     def __len__(self):
         return len(self.imgs)
-
-
-if __name__ == '__main__':
-    Synthetic("trial", dummy_gen, 10)
